Add core pair helpers and fix ABI types

+++ src/interfaces/IPair.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.30;
 
 interface IPair {
-    function getReserves() external view returns (uint256 reserve0, uint256 reserve1, uint256 lastTimeStamp);
+    function getReserves() external view returns (uint112 reserve0, uint112 reserve1, uint32 lastTimeStamp);
     function swap(uint256 amount0Out, uint256 amount1Out, address to, bytes calldata data) external;
     function initialize(address _token0, address _token1) external;
     function mint(address to) external returns (uint256 liquidity);
